Here's the stuff I got so far
Basically a bunch of place holders. The homepage is just a general idea of what we'd present to the user, and we'll add links to everything later once we figure out where we're gonna host this.
The instruction manual has nothing in it. I figure it'll be a copy/paste of our submitted user manual??
Local attractions has nothing in it. It'll get the good feed stuff integrated in later, once we figure out how we want to design the project and present everything to the user.
The settings has a couple of sliders, but I didn't know what else to add, since a lot of our stuff will probably be map filters. Maybe if we add more later, there will be more things to populate it with?